+++ app/panelists/reports/average_scores_by_year.py
@@ -4,6 +4,7 @@
 #
 # vim: set noai syntax=python ts=4 sw=4:
 """WWDTM Panelists Average Scores Report Functions."""
+
 from decimal import Decimal
 
 from flask import current_app