+++ app/panelists/reports/common.py
@@ -4,6 +4,7 @@
 #
 # vim: set noai syntax=python ts=4 sw=4:
 """WWDTM Common Panelist Report Functions."""
+
 from typing import Any
 
 from mysql.connector.connection import MySQLConnection