+++ app/shows/reports/guest_host_scorekeeper.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 # vim: set noai syntax=python ts=4 sw=4:
-"""WWDTM Show Guest Hosts Report Functions."""
+"""WWDTM Show Guest Host and Scorekeeper Report Functions."""
 
 from mysql.connector.connection import MySQLConnection
 from mysql.connector.pooling import PooledMySQLConnection
@@ -11,17 +11,18 @@ from mysql.connector.pooling import PooledMySQLConnection
 from app.shows.reports import show_details
 
 
-def retrieve_shows_guest_host(
+def retrieve_shows_guest_host_scorekeeper(
     database_connection: MySQLConnection | PooledMySQLConnection,
 ) -> list[dict]:
-    """Retrieve a list of shows with guest hosts."""
+    """Retrieve a list of shows with a guest host and guest scorekeeper."""
     if not database_connection.is_connected():
         database_connection.reconnect()
 
     query = """
         SELECT s.showid, s.showdate, s.bestof, s.repeatshowid, h.host,
-        h.hostslug, sk.scorekeeper, sk.scorekeeperslug,
-        skm.guest as scorekeeper_guest, l.venue, l.city, l.state
+        h.hostslug, hm.guest AS host_guest, sk.scorekeeper,
+        sk.scorekeeperslug, skm.guest AS scorekeeper_guest, l.venue,
+        l.city, l.state
         FROM ww_showhostmap hm
         JOIN ww_hosts h ON h.hostid = hm.hostid
         JOIN ww_showskmap skm ON skm.showid = hm.showid
@@ -29,7 +30,7 @@ def retrieve_shows_guest_host(
         JOIN ww_shows s ON s.showid = hm.showid
         JOIN ww_showlocationmap lm ON lm.showid = hm.showid
         JOIN ww_locations l ON l.locationid = lm.locationid
-        WHERE hm.guest = 1
+        WHERE hm.guest = 1 AND skm.guest = 1
         ORDER BY s.showdate ASC;
         """
     cursor = database_connection.cursor(dictionary=True)
@@ -54,6 +55,7 @@ def retrieve_shows_guest_host(
                 },
                 "host": row["host"],
                 "host_slug": row["hostslug"],
+                "host_guest": row["host_guest"],
                 "scorekeeper": row["scorekeeper"],
                 "scorekeeper_slug": row["scorekeeperslug"],
                 "scorekeeper_guest": bool(row["scorekeeper_guest"]),