+++ app/shows/reports/guests_vs_bluffs.py
@@ -4,6 +4,7 @@
 #
 # vim: set noai syntax=python ts=4 sw=4:
 """WWDTM Show Not My Job Guest Wins Rate vs Bluff the Listener Wins Rate Report Functions."""
+
 from typing import Any
 
 from mysql.connector.connection import MySQLConnection