+++ app/shows/reports/search_multiple_panelists.py
@@ -4,6 +4,7 @@
 #
 # vim: set noai syntax=python ts=4 sw=4:
 """WWDTM Search Shows by Multiple Selected Panelists Report Functions."""
+
 from typing import Any
 
 from mysql.connector.connection import MySQLConnection