+++ tests/test_locations.py
@@ -4,6 +4,7 @@
 #
 # vim: set noai syntax=python ts=4 sw=4:
 """Testing Locations Module and Blueprint Views."""
+
 from flask.testing import FlaskClient
 from werkzeug.test import TestResponse
 
@@ -14,6 +15,16 @@ def test_index(client: FlaskClient) -> None:
     assert response.status_code == 200
     assert b"Locations" in response.data
     assert b"Average Scores by Location" in response.data
+
+
+def test_home_vs_away(client: FlaskClient) -> None:
+    """Testing locations.routes.home_vs_away."""
+    response: TestResponse = client.get("/locations/home-vs-away")
+    assert response.status_code == 200
+    assert b"Home vs Away" in response.data
+    assert b"Year" in response.data
+    assert b"Home" in response.data
+    assert b"Away" in response.data
 
 
 def test_average_scores_by_location(client: FlaskClient) -> None: